dont assume there is only 1 new connection on server socket event
instead, call accept in a loop until wouldblock is returned.
Before, we'd just call it once after an event, which would cause
connections not to be accepted, and eventually the event is not
fired at all because we use edge triggered epoll, so no connection
was accepted anymore. this fixes that.