Fix panic at end of file download
After transitioning from Reading to Ready/EndReached state, we increased the
block offset for the next read. Before calling the finish_read method we
would keep in a local variable the block offset to use it in range calculations to trim off
parts of the last read block if it fell out of the requested range.

This caused a bug where you would call try_get_read_bytes multiple times on
the request before the last one, it would return you the intended length for
the last request on the second call. This caused a panic in FileResponder
because it needs to use the same buffer multiple times since the socket
send buffer is smaller than the file read buffer.

This fixes the bug, and adds a regression test.