refactor file Reader + fix small issue + optimization
The state machine inside file Reader was written differently than
the pattern used elsewhere (Option<State>) and was hard to understand.
It also contained some hacks (to keep old block_index) and would read more
than needed on the last request in most cases.

The request range tracking code has been moved to another struct, that just
gives an Option<new range> whenever the next read operation is started and
has a method to translate to buffer range (ignoring the parts of the buffer
that do not overlap with the requested total range passed to the reader)
and the operation range (block aligned and multiple block length).

Also, there was (likely, test also contained error) a bug with
block aligned total ranges, an extra request would be queued
and a buffer of length 0 returned. This is also fixed with this refactoring.