work on having socket being owned by connection instead of handlers
the tls layer will have to provide a wrapper to the socket
so de/encryption is transparant to the http & app layer.

the tls socket layer contains big structs and is not an ideal
candidate for moving the socket from one owner to the next as we
do now.

also, to both support a normal TcpStream and a DecryptedSocket
we need to use runtime polymorphish.

A solution for both problems could be to use Box<Trait>, but we'd
rather not use the heap. So we need to use trait objects, and thus ownership
needs to be transfered to a more central location: the connection struct

As it is, this doesn't compile because of lifetime mismatches.