fix socket wrapping compilation errors
to create a context with a wrapped socket, we need a splitting borrow on the context
for this, I created the ContextFactory which can be turned back into a Context
given a (new) socket. The fields (apart from socket) of Context are now repeated which
isn't great. The alternative would be to have something like ContextBase that is used in both
but we'd need Context to have a reference to it, so in server.rs we'd have to create the base
first and then the context so the base has storage on the stack for the reference held in context.

this works as well, if the fields repetition turns out to be annoying, we can always switch