don't support pushing multiple buffers to skey_decoder for now
I don't know how to model this with rust lifetimes, and we don't need it for now. The private key is backed by memory (which in case of multiple pushes? all buffers? which lifetime is that?) from the buffer(s), but the decoder we only need temporarily, but the resulting Key should have the correct lifetime.