implement tls connection closing
before the tcp socket can be closed, we need to send a close_notify alert at the
tls level to avoid truncation attacks. The other end will also respond with an alert,
so we need to close the tls connection async, meaning we need state in the tls handler (is_closing).
Once this exchange of alerts is finished and the bearssl engine marks itself as closed,
we close the tcp socket by reporting termination to the parent in handle_event.

Some code duplication remains in the commit, needs to be cleaned up in a subsequent commit.